Map Decodable error to APIGeneratorError
Why:
The APIGeneratorError type provides simple error messages when compared
to the Decodable error message for a missing key. Leaving the raw
Decodable error also leaks internal details that can be confusing to
users of the tool.

This change addresses the need by:
Capture `keyNotFound` errors and map them to a `buildArgumentRequired`.