Improve reporting for missing deployment target
Why:
BuildArguments uses a force unwrap when accessing `deploymentTarget`.
The access of this property and the creation of a `BuildArguments`
value could be far apart in the code, which might make it harder to
diagnose why the code crashes.

This change addresses the need by:
Proactively check to see if accessing `deploymentTarget` will crash at
the point a `BuildArguments` value is constructed and raise an error
that is reported more gracefully.