Make TextProcessor Protocol more composable
Why:
Having do/catch blocks means we have to re-map
the errors making the call-site look a litte odd.
The result type offers a cleaner way to represent
the value and the error, which in this case is
already wrapped up in the type we need.

This change:
Use Result over do/catch for TextProcessor