Remove the ability to write to a file
Why:
It's far much easier to let the user redirect stdout to a file instead.
Can remove some error handling from the app.
There might be usecases where the end user woould like to use the output
of guise in a different tool.

This change addresses the need by:
Remove writeToFile and replace with fputs

Fix #10